(refactor) remove database package
